refactor: reorganize welcome page components and consolidate animation logic
