+++ src/hooks/useResponsiveModel.js
@@ -13,7 +13,7 @@ export function useResponsiveModel() {
       } else if (width < 900) {
         setModelScale(1);
       } else {
-        setModelScale(1.4);
+        setModelScale(1.8);
       }
 
       if (height < 700) {
