+++ src/welcomePage/Welcome.css
@@ -2,24 +2,47 @@
 
 .welcome-container {
   position: fixed;
-  top: 3vh;
+  top: 0;
   left: 0;
   z-index: 2;
   width: 100%;
   height: 100%;
   display: flex;
-  justify-content: space-between;
-  padding: 3rem 2rem 3rem 2rem;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 2rem;
   font-family: 'IBM Plex Mono', monospace;
   letter-spacing: 0.1em;
   font-weight: 600;
   color: black;
   pointer-events: none;
   z-index: 10;
-  align-items: flex-start;
   box-sizing: border-box;
   overflow: hidden;
   overscroll-behavior: none;
+  background-color: #ffffff;
+  background-image: radial-gradient(#13121279 1px, transparent 1px);
+  background-size: 10px 10px;
+  gap: 2rem;
+}
+
+.welcome-container::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  z-index: -1;
+  background: linear-gradient(to bottom, rgba(255,255,255,1), rgba(255,255,255,0.8));
+  pointer-events: none;
+}
+
+.welcome-header-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  width: 100%;
+  padding: 2.5rem;
+  
 }
 
 .welcome-left {
@@ -27,8 +50,7 @@
   font-size: 0.9rem;
   line-height: 1.1;
   white-space: pre-wrap;
-  margin: 0 auto;
-  flex: 1 1 0;
+  flex: 1;
 }
 
 .intro-heading {
@@ -43,8 +65,7 @@
   line-height: 0.7;
   gap: 0.3rem;
   align-items: flex-end;
-  flex: 1 1 0;
-  margin: 0 auto;
+  flex: 1;
 }
 
 .label-block {
@@ -89,69 +110,14 @@
 .see-more-link:hover {
   text-decoration: underline;
 }
-
-@media (max-width: 900px) {
-.project-info img {
-  
-  max-width: 4rem;
-  max-height: 4rem;
-  height: auto;
-  border: 1px solid black;
-  border-radius: 6px;
-  margin-bottom: 0.6rem;
-  margin-left: auto;
-  margin-right: auto;
-}
-.project-info {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
+.djscene-canvas {
+  background: transparent;
   position: fixed;
-  bottom: 5rem;
-  gap: 1.2rem;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: rgba(255, 255, 255, 0.85);
-  padding: 0.8rem;
-  width: 80vw;
-  max-width: none;
-  font-size: 10px;
-  border-radius: 8px;
-  
-  z-index: 10;
+  width: 100vw;
+  height: 100vh;
+  z-index: 1;
+  overflow: hidden;
+  touch-action: none;
+  overscroll-behavior: none;
   pointer-events: auto;
-  border: 3px solid rgb(0, 0, 0);
-}
-
-
-}
-
-@media (max-width: 600px) {
-  .welcome-container {
-    top: 1vh;
-    font-size: 0.82rem;
-    padding: 3rem 0.4rem 0rem 0.6rem;
-  }
-  .welcome-left {
-  max-width: 600px;
-  font-size: 0.7rem;
-  line-height: 1.1;
-  white-space: pre-wrap;
-  margin: 0 0;
-  
-  flex: 1 1 0;
-}
-
-.welcome-right {
-  text-align: right;
-  display: flex;
-  flex-direction: column;
-  font-size: 0.6rem;
-  line-height: 0.7;
-  gap: 0.3rem;
-  align-items: flex-end;
-  flex: 1 1 0;
-  margin: 0 auto;
-}
 }
