+++ src/welcomePage/components/AnimatedButton.jsx
