+++ src/welcomePage/components/SpaceButton.css
@@ -23,6 +23,7 @@
   flex-direction: column;
   align-items: center;
   z-index: 1000;
+  pointer-events: auto;
 }
